Apply ruff rule RUF046
RUF046 Value being cast to `int` is already an integer

+++ ome_zarr/csv.py
@@ -20,7 +20,7 @@ def parse_csv_value(value: str, col_type: str) -> Union[str, float, int, bool]:
         if col_type == "d":
             rv = float(value)
         elif col_type == "l":
-            rv = int(round(float(value)))
+            rv = round(float(value))
         elif col_type == "b":
             rv = bool(value)
     except ValueError:
